imported osmosis libraries

+++ src/sasa_importer/street_network/components/DenseInfo.java
@@ -5,10 +5,10 @@ public class DenseInfo {
 	int version;
 	long timestamp;
 	long changeset;
-	long latitude;
-	long lognitude;
+	double latitude;
+	double lognitude;
 	
-	public DenseInfo(int version, long timestamp, long changeset, long latitude, long lognitude) {
+	public DenseInfo(int version, long timestamp, long changeset, double latitude, double lognitude) {
 		this.version = version;
 		this.timestamp = timestamp;
 		this.changeset = changeset;
@@ -40,19 +40,19 @@ public class DenseInfo {
 		this.changeset = changeset;
 	}
 	
-	public long getLatitude() {
+	public double getLatitude() {
 		return latitude;
 	}
 	
-	public void setLatitude(long latitude) {
+	public void setLatitude(double latitude) {
 		this.latitude = latitude;
 	}
 	
-	public long getLognitude() {
+	public double getLognitude() {
 		return lognitude;
 	}
 	
-	public void setLognitude(long lognitude) {
+	public void setLognitude(double lognitude) {
 		this.lognitude = lognitude;
 	}
